+++ src/main/java/com/peppone/dam/board/domain/BoardType.java
@@ -1,0 +1,7 @@
+package com.peppone.dam.board.domain;
+
+public enum BoardType {
+  BOARD_TYPE_MINI,
+  BOARD_TYPE_MINOR,
+  BOARD_TYPE_MAJOR
+}
