+++ src/main/java/com/peppone/dam/board/service/BoardService.java
@@ -1,0 +1,10 @@
+package com.peppone.dam.board.service;
+
+import com.peppone.dam.user.domain.UserEntity;
+import com.peppone.dam.board.dto.BoardMakingDto;
+import com.peppone.dam.response.CommonResponse;
+
+public interface BoardService {
+
+  CommonResponse makeBoard(BoardMakingDto boardMakingDto, UserEntity user);
+}
