--- src/main/java/com/peppone/dam/domain/BoardEntity.java
@@ -1,43 +1,0 @@
-package com.peppone.dam.domain;
-
-import jakarta.persistence.ElementCollection;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.CollectionId;
-
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-@Table(name = "board")
-@Builder
-public class BoardEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private long id;
-
-  private String name;
-
-  private String url;
-
-  @Enumerated(EnumType.STRING)
-  private BoardType boardType;
-
-  @ManyToOne
-  @JoinColumn(name = "admin_id")
-  private UserEntity userEntity;
-}
