--- src/main/java/com/peppone/dam/domain/BoardType.java
@@ -1,7 +1,0 @@
-package com.peppone.dam.domain;
-
-public enum BoardType {
-  BOARD_TYPE_MINI,
-  BOARD_TYPE_MINOR,
-  BOARD_TYPE_MAJOR
-}
