--- src/main/java/com/peppone/dam/domain/PostEntity.java
@@ -1,45 +1,0 @@
-package com.peppone.dam.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-@Table(name = "posts")
-@Builder
-public class PostEntity {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private long id;
-  @ManyToOne
-  @JoinColumn(name = "user_id")
-  private UserEntity userId;
-  @ManyToOne
-  @JoinColumn(name = "board_id")
-  private BoardEntity boardId;
-  private String subject;
-  private String content;
-  private LocalDateTime createdTime;
-  private LocalDateTime modifiedTime;
-  private LocalDateTime deletedTime;
-  private long like;
-  private long dislike;
-  private double likeRatio;
-  private boolean access;
-
-}
