+++ src/main/java/com/peppone/dam/post/domain/PostType.java
@@ -1,0 +1,7 @@
+package com.peppone.dam.post.domain;
+
+public enum PostType {
+  POST_TYPE_GENERAL,
+  POST_TYPE_PINNED,
+  POST_TYPE_NOTICE
+}
