+++ src/main/java/com/peppone/dam/post/repository/PostRepository.java
@@ -1,0 +1,8 @@
+package com.peppone.dam.post.repository;
+
+import com.peppone.dam.post.domain.PostEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<PostEntity, Long> {
+
+}
