--- src/main/java/com/peppone/dam/repository/PostRepository.java
@@ -1,8 +1,0 @@
-package com.peppone.dam.repository;
-
-import com.peppone.dam.domain.PostEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<PostEntity, Long> {
-
-}
