--- src/main/java/com/peppone/dam/service/PostService.java
@@ -1,10 +1,0 @@
-package com.peppone.dam.service;
-
-import com.peppone.dam.domain.UserEntity;
-import com.peppone.dam.dto.CreatePostDto;
-import com.peppone.dam.response.CommonResponse;
-
-public interface PostService {
-
-  CommonResponse createPost(UserEntity user, CreatePostDto createPostDto);
-}
