--- src/main/java/com/peppone/dam/service/TokenService.java
@@ -1,8 +1,0 @@
-package com.peppone.dam.service;
-
-import com.peppone.dam.domain.UserEntity;
-
-public interface TokenService {
-  UserEntity tokenValidation(String token);
-  String tokenIssuer(UserEntity user);
-}
