+++ src/main/java/com/peppone/dam/user/domain/UserEntity.java
@@ -1,0 +1,90 @@
+package com.peppone.dam.user.domain;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonProperty.Access;
+import jakarta.persistence.ElementCollection;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.stream.Collectors;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+@Table(name = "user")
+@Builder
+public class UserEntity implements UserDetails {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long id;
+
+  private String userEmail;
+
+  private String password;
+
+  private String nickname;
+
+  private LocalDateTime createdDate;
+
+  private LocalDateTime modifiedDate;
+
+  private LocalDateTime removedDate;
+  @ElementCollection(fetch = FetchType.EAGER)
+  @Builder.Default
+  private List<String> role = new ArrayList<>();
+
+  private int block_day;
+
+  @Override
+  public Collection<? extends GrantedAuthority> getAuthorities() {
+    return this.role.stream().map(SimpleGrantedAuthority::new).collect(Collectors.toList());
+  }
+
+  @JsonProperty(access = Access.WRITE_ONLY)
+  @Override
+  public String getUsername() {
+    return this.userEmail;
+  }
+
+  @JsonProperty(access = Access.WRITE_ONLY)
+  @Override
+  public boolean isAccountNonExpired() {
+    return true;
+  }
+
+  @JsonProperty(access = Access.WRITE_ONLY)
+  @Override
+  public boolean isAccountNonLocked() {
+    return true;
+  }
+
+  @JsonProperty(access = Access.WRITE_ONLY)
+  @Override
+  public boolean isCredentialsNonExpired() {
+    return true;
+  }
+
+  @JsonProperty(access = Access.WRITE_ONLY)
+  @Override
+  public boolean isEnabled() {
+    return true;
+  }
+}
