+++ src/main/java/com/peppone/dam/user/repository/UserRepository.java
@@ -1,0 +1,15 @@
+package com.peppone.dam.user.repository;
+
+import com.peppone.dam.user.domain.UserEntity;
+import java.util.List;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
+  UserEntity findByUserEmail(String email);
+
+  List<UserEntity> findAll();
+
+  UserEntity findById(Optional<UserEntity> byId);
+}
