+++ src/main/resources/application.properties
@@ -11,5 +11,4 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.hibernate.ddl-auto=create
-
 spring.jpa.properties.hibernate.auto_quote_keyword=true
